Simplify ternary byte casts
No longer needed to explicitly cast, thanks c#9 !

+++ PKHeX.Core/Editing/CommonEdits.cs
@@ -246,7 +246,7 @@ namespace PKHeX.Core
             if (pk is IGigantamax c)
                 c.CanGigantamax = Set.CanGigantamax;
             if (pk is IDynamaxLevel d)
-                d.DynamaxLevel = (byte)(d.CanHaveDynamaxLevel(pk) ? 10 : 0);
+                d.DynamaxLevel = d.CanHaveDynamaxLevel(pk) ? 10 : 0;
 
             pk.ClearRecordFlags();
             pk.SetRecordFlags(Set.Moves);
